requirements.txt
asynctest==0.13.0
nose2==0.10.0
Pillow==9.0.0
psycopg2-binary==2.9.2
pyquery==1.4.3
pytz==2021.3
telepot==12.7
